Make it easier to run the website
Summary:
Adding a couple commands to package.json:
- `yarn start-internal-docs`
- `yarn start-with-wasm`
- `yarn start-internal-docs-with-wasm`
- `yarn build-with-wasm`
- `yarn install-with-wasm-deps`

This makes it so that the developer don't need to worry about directly running any of the shell scripts. This also set it up so that the workflow simply becomes:
- `yarn install-with-wasm-deps` when setting up the project

Run one of the `yarn start` command depending on whether or not they want to build the internal site and if they want to rebuild wasm. There's also `yarn build` commands if the developer only wants to run build oncce and rely on `yarn serve`

Reviewed By: kinto0

Differential Revision: D73527318

fbshipit-source-id: 6487e1839a5998d75ac1e9b352c970c7d6c3d824