Add a test of descriptors on awaitable funcs
Summary:
I am seeing some errors in an example project that look
as though wrapping async functions in decorators does not
reliably work. It seems to be okay based on this test, I
think it's worth having this as a test case even though it
does not reveal a bug.

Reviewed By: rchen152

Differential Revision: D73522693

fbshipit-source-id: 0acb9d92e60112c07ea5a04575eccb9d19a6a1dc