Add test cases for class-object `__getattr__`
Summary:
When we try to access an attribute on a class object, the runtime does not
try to use the `__getattr__` method defined on the class - it will only use
that for *instance* lookups.

Instead, it tries to use the metaclass `__getattr__`.

Pyrefly's current lookup logic produces two bugs here: first, we try to use
the `__getattr__` defined on the class, which leads to call errors since
it is an unbound method when accessed on the class itself. Second, the
class `__getattr__`, if defined, will *hide* the metaclass `__getattr__`,
thereby producing a large number of type errors and potentially leading
to bad type inference if the two don't return the same type.

Note that in the case where the class *doesn't* define `__getattr__`, the
dispatch against the metaclass is working fine, it's only when that is
shadowed that we get bad behavior.

Reviewed By: rchen152

Differential Revision: D73522691

fbshipit-source-id: c04d45ab69d21349ed09c8325d30ad7b5d24af61