Add Introduction page
Summary:
Adding an introduction page to the docs. This will be the landing spot when the user first click on "Docs" in the navbar or the landing page header. This was previously just a table of content for the docs. I moved what I was adding for the Getting Started page in D73453913 here since the content fits much better under "Introduction" rather than "getting started".

We also rename "Introduction" category to "Getting Started" and have that only include "Installtion" and "Configuration"

Reviewed By: MaggieMoss

Differential Revision: D73533701

fbshipit-source-id: 6cf937081586c53cf2a007430f76f06ba9cb61ad