First cut of local find references: exposed as document highlight
Summary:
This diff is the first step to implement global find references. In general, we will follow the following strategy:

1. Go to the definition
2. For each file that can possibly contain the definition, we  try to find all variables/attributes that has the definition.
3. Combine the result

There are still some work to do to make 2 performant and non-blocking, but let's first have the skeleton code that can find all the references locally, so that later we can reuse the same code for global references.

This diff implements the first cut of local references: we first decide whether the definition is a variable or attribute. If it's variable, then we look for all usages in bindings. If it's attribute, then we look check every  attribute access in the file to see whether the attribute has the same definition location as the definition.

Finally, we will expose the implementation as `textDocument/documentHighlight`

Reviewed By: kinto0

Differential Revision: D73529881

fbshipit-source-id: bfeb8188d6abefa1e3a5262fa816f3ece4599eaf