Fix broken anchor link, partially fix another
Summary:
I am unable to link to "Stub Files vs Source Files" successfully
no matter what I do, the docs suggest that an automated id should
exist and that we should be able to set an explicit id, see
https://docusaurus.io/docs/markdown-features/toc, but both approaches
still give me a broken anchor link warning.

But at least the page portion of the link now works, before the link
was actually completely broken (and didn't fail the build, because
docusaurus considers a broken anchor link a warning even if the page
part is a dead link).

Reviewed By: dluo

Differential Revision: D73603399

fbshipit-source-id: 0c9a15db268ccf2d86c8b17a12cc0df63016a0bd