fix last broken anchor link
Summary:
Docusaurus doesn't play well with docs that uses headers (`#`). Changing this to `##` fixes the error.
This seem to be by design based on the following issue: https://github.com/facebook/docusaurus/issues/5036. We should make sure to not use h1 heading anchors aside from the top of the page and should never link to them.

Reviewed By: connernilsen

Differential Revision: D73606880

fbshipit-source-id: 78cd2e5ae516c8cd3520d42286140cbda476ea0c