Add module path to class and qname equality
Summary: Otherwise we get a crash with the test plan below, because we end up with two typing.py files in scope, and they don't agree with each other. I think this was always wrong, but previously we mostly had the mixed in one winning, and didn't get a crash. Now we do get a crash, but can fix that by properly using path as the key.

Reviewed By: kinto0

Differential Revision: D73586571

fbshipit-source-id: 35575486ac8ee20250ad6a3c0c15f71ec33c2154