Canonicalize `typing.Type` into `type[any]` in untype
Summary:
When the bare name `typing.Type` is used as annotation, it's expected to behave like `type[any]`. This is currently not implemented yet, and this diff attempts to add the special-casing.

I had lot of troubles deciding where to add it. Right now I put it next to where the `typing.Tuple` -> `tuple[Any, ...]` expansion lives. I'm actually not 100% sure this is the right place, since the expansion won't actually happen on non-annotation positions -- but that seems to be exactly how mypy and pyright behaves... So I think we could live with it for now.

The other idea is to add the special case to `subset.rs`, where we just handle `typing.Type` within the subset eq check. This is actually what happened to `typing.Type`'s brother -- the builtin `type` type. But my current understanding is that we had this implicit assumption that certain kinds of types, like SpecialForm or SuperInstance, is expected to be “desugared” outside of the solver before the subset_eq queries get hit.

Reviewed By: ndmitchell

Differential Revision: D73563384

fbshipit-source-id: 74006f8329fcdb85438f4abc1b1b1bdc9c529bce