Simplify attribute checking
Summary:
Move attribute checking / narrowed type inference entirely into `attr.rs` instead
of having it split into a layer in `solve.rs` and a layer in `attr.rs`. This makes it
easier to see how the top-level "value might be an expr or might be a binding / type"
logic propagates down to the meat of the code inside `attr.rs`, and it gets rid of two
helper functions that weren't really all that necessary, cutting out one layer of
abstraction.

Reviewed By: rchen152

Differential Revision: D73600004

fbshipit-source-id: 86fecda815cd57786081e7563308c0d051416525