Make attribute assignment bind
Summary:
An attribute assignment *may* bind the top-level name, for example
`x.y.z = 5` will bind `x` with narrowing applied to the `y.z`
attribute chain.

This commit wires up that behavior - which we had already prepped
the solve-side code for by just having it forward the existing type
info of `x` for now - in the bindings code. We use `Key::AttrAssign`
to make it clearer from the key type what is going on; in principle
a `Key::Definition` would work fine but that seems misleading.

Note that not all attribute assignments actually bind a name; if the
base is anonymous, as in `f().y.z = 5`, then no actual binding will
occur; we just use `Key::Anon` for this case.

Reviewed By: rchen152

Differential Revision: D73599994

fbshipit-source-id: 21ef0504b4de1f8ec41e20cc5b1a64867c7536b6