Implement attribute narrowing x assignment logic
Summary:
This is the key diff we've been working toward: now that assignments to attributes
where the base is a name or an attribute chain can bind the top-level name, and
attribute checking also returns an optional narrow type based on inference of the
value assigned, we can glue these together to update the TypeInfo on assignment.

The update handles both invalidating any existing subtree under the assigned attribute
chain (unconditionally), and (assuming a narrowed type is available) narrowing
to reflect the assignment.

------

Note: In code with a lot of attribute assignments, this is going to cause the
TypeInfo size to grow a bit. (I think it's probably much more common to have a lot
of assignments than a lot of narrowing conditions - assignments occur in linear
control flow whereas conditions normally don't.)

If we discover that this is a problem, there's a space optimization I'm not
doing, which is that if we could determine the attribute *read* type exactly
matches the narrowed type, we could skip producing a narrow.

That would cost compute (every attribute set would also have to do an attribute
read), but would save RAM. ndmitchell I'm curious your opinion - the logic to
make that optimization belongs in `attr.rs`; I could stack some more work to do
it, either now or after PyCon.

Reviewed By: rchen152

Differential Revision: D73599998

fbshipit-source-id: 76ae691d7784e8d7b5037248d1f9da4235a944d6