Add `bind_attr_assign` helper
Summary:
The logic here needs to be more involved than it currently is, where we'll actually
bind the name being assigned when there is one.

As a result, it seems worth deduplicating the logic since the current trivial call
to `self.table.insert` will become a small block with control flow.

Reviewed By: rchen152

Differential Revision: D73600001

fbshipit-source-id: 33a0b827db23e3770fc17bb7f63ae1056de35897