Make some plumbing changes to prep for handling assignment
Summary:
This commit begins re-wiring Pyrefly plumbing so that we can eventually
treat attribute assignments as potentially impacting attribute narrows.

This is pure type tetris; we are not yet doing the bindings-side work; the type
info is just a placeholder for now, and we aren't doing the work to actually
use it in bindings.

NOTE: not all attribute assignments will actually narrow. It's possible to assign to
an anonymous base, e.g. `f().x = 42`. My current plan is to just use a `Binding` for
assignment in all cases and have the resulting `TypeInfo` not propagate anywhere; it's
easier to keep the logic all in one place this way even though the types aren't quite
as informative. I can potentially go back and undo this later if there are objections,
but I'd like to at least get it working using the simpler approach first.

Reviewed By: rchen152

Differential Revision: D73599993

fbshipit-source-id: 37673e1136e0cc391c917f4baa84d5bda804fdf6