Add descriptor narrowing test, marked as bug
Summary:
The descriptor narrowing logic works as intended (writes do not narrow, because
that's not possible to do in a way that is both sane and simple), but is probably
not good enough for the long-term.

This test case verifies the current behavior and also describes where we likely
want changes.

Reviewed By: rchen152

Differential Revision: D73599999

fbshipit-source-id: 64fa59cc7c002f317c781637c8c3bcb95a4d8d85