Add logic to propagate TypeInfo on attribute assign
Summary:
In order to handle both adding and invalidating narrows for attribute assignment, we
need to have enough plumbing in place to look up the existing top-level name.

This requires looking at the assignment target and *optionally* grabbing an Identifier
for it (which can then be used to construct the `Key::Usage` needed for looking up existing
type info). The lookup is optional because not all assignment targets have a top-level name,
it is possible to use an anonymous target like `f().x = 42)` in which case we'll get `None`
in the lookup (and the resulting `TypeInfo` will be ignored).

With this in place, we are now ready to work on the bindings-side logic (if we had tackled the
bindings-side before wiring up propagation, we'd break existing behavior). We'll add the actual
narrowing logic afterward, once the code is testable.

Reviewed By: rchen152

Differential Revision: D73599996

fbshipit-source-id: 5e08c2003649527eb75f5abe81c703c607af3f3d