Implement pyproject parsing for mypy configs
Summary:
Mypy allows its pyproject.toml configs to be backwards compatible (in a sense) with the mypy.ini configs. This basically means that every entry can be a single string, a list of strings, or a string that needs to be split into a list.

For example, `mypy_path` is documented as a single string that may contain multiple paths that are delimited by `:` or `,`, but mypy allows this to be a list of strings in the TOML file.

This means it's hard to reuse code between the INI parser and the TOML parser right now. Rewriting the INI parser may make it much easier to share code, and this is left for a future refactor.

Next diffs:
- pyproject.toml parsing for pyright
- wire up pyproject.toml parsing to the command itself

Reviewed By: connernilsen

Differential Revision: D73613371

fbshipit-source-id: 71f63cfa2c375b94523bfc379f5e267c1e79b25c