Rename `type_of_attr_get_if_found` and `call_method` to make it clear they only ever get invoked for magic dunder methods
Summary:
whats_in_a_name

If we look closely, these two methods are only ever invoked for *magic dunder* attributes only, even if the func name seems to suggest that they'd work for all kinds of attributes. Magic dunder attributes are those dunder attributes where if being looked up from a base of `type[A]`, we search for the attribute within the definition of `type` instead of the definition of `A`. Therefore they follow different lookup rule from other attributes.

In D73678063 we are going to leverage such observation and restrict the logic to only work for magic dunder attributes. Therefore let's rename these funcs proactively to make it clear what they do. Note that In the past, the need for special-casing magic dunder attributes has been independently observed in multiple occasions by multiple people (see D73537185 and D73293425). But the solutions are implemented separately in similar but not identical ways. So this stack will also attempt to provide a shared solution and unify those downstream implementations.

NOTE: The name "magic dunder attribute" is not an "official" terminology. It's a term coined by Steven to make it easier to talk about the attr lookup behaviors exhibited by these attributes. The Python runtime doesn't really store these attributes in any special way -- the special behaviors are implemented by simply swapping out the function pointers on the `type` object for the corresponding methods that depend on those attrs.

Reviewed By: stroxler

Differential Revision: D73678065

fbshipit-source-id: 6719d589726efc11a561f7ce92e4a81e29de3c69