Bump patch version
Summary:
Two days ago, we restructured the typeshed directory used in Pyrefly's
build.

This means it is not possible to run mypy_primer without a lot of extra
flags, because the defualt behavior is to compare the latest release vs
trunk for a typechecker, and one or the other is going to panic (because
the build scripts are coupled to the directory layout).

Triggering a patch version bump because there are no major changes, we
just need a tag to exist so mypy_primer can find it.

Reviewed By: rchen152

Differential Revision: D73692380

fbshipit-source-id: 549f9df493f0417e10da7e855f4c6367fbcfe229