Only show buttons on code snippets on hover
Summary:
The new buttons takes up a good amount of space. It'll be better if they are hidden by default, and shows up only on hover. I got this idea from the rust docs (example: https://doc.rust-lang.org/book/ch13-01-closures.html).

Note that on mobile, we choose to always show the button since it's much harder to hover on the code snippet

Reviewed By: stroxler

Differential Revision: D73609944

fbshipit-source-id: b6d1259c8d34c8c9a8c4af126ece7474113406b6