symlink rust-toolchain file from pyrefly/
Summary:
not everyone runs our internal setup_cargo.sh script. if it's not run, rust will use the toolchain from `rust-toolchain` file in pyrefly/pyrefly. we should use the same one in the wasm build.

symlinking this is not the best solution: i'm not sure if this will work on windows / clone to git correctly. I'd rather do something like D73783220. But D73783220 doesn't work in it's current form and I don't understand `wasm-pack` enough to get it to work or contribute upstream

Reviewed By: dluo

Differential Revision: D73784919

fbshipit-source-id: 56ab2486b2038007d7b1cd34358117f7c6e46d7a