Populate all files from `--experimental_project-path`
Summary:
For find references to work, we need a full picture of the entire project the user cares about to know the reverse dependency map.

In this diff, I first add `experimental_project-path` config arg to lsp command, as a temporary measure to give us the information. In the future, the knowledge should come from the config.

Then I will check all the files under that directory. It could be made into a background job. However, it can cause non-determinism in tests, and it shouldn't be too bad for most projects, so I kept it that way.

Finally, I gated the `referencesProvider` declaration behind having something in `--experimental_project-path`, effectively gating find-references behind this flag.

Reviewed By: connernilsen

Differential Revision: D73632434

fbshipit-source-id: 4a562601198e507dbddd54cc3d3e2c430b21e13d