Add unit test of NamedTuple qualifier bug
Summary:
A NamedTuple cannot use `ClassVar` or `Final` qualifiers in members, but the resulting
class can be used as a bass class, and subclasses are permitted to have qualifiers
in their fields.

We understand the subclass fields just fine, but the code to catch qualifiers is
incorrectly marking them as a problem.

Reviewed By: ndmitchell

Differential Revision: D73854188

fbshipit-source-id: 487c3ec6fa34717d37f6ebd68b913ecbc1fe6d8c