Fix the error in `enum.pyi` for < 3.11
Summary:
We just need to special case one more descriptor class. It didn't come up in earlier
tests because the stub file has
```
# in enum.pyi
if sys.version_info >= (3, 11):
    # In 3.11 `EnumMeta` metaclass is renamed to `EnumType`, but old name also exists.
    EnumType = EnumMeta

    class property(types.DynamicClassAttribute):
        def __set_name__(self, ownerclass: type[Enum], name: str) -> None: ...
        name: str
        clsname: str
        member: Enum | None

    _magic_enum_attr = property
else:
    _magic_enum_attr = types.DynamicClassAttribute
```
and we special cased `enum.property` already, but not `types.DynamicClassAttribute`.

This fixes half of issue 137 (qualified attributes of NamedTuple are a separate problem).

Reviewed By: ndmitchell

Differential Revision: D73851724

fbshipit-source-id: 5791c8ccc67cd88ec05e5c30ea3a23904fbd08a4