Set use_untyped_imports to be true by default
Summary:
Aaron mentioned that we should try to work by default, and Neil agreed that `use_untyped_imports` should probably not fail unless specified.

This diff updates the behavior to `use_untyped_imports` by default.

I was also looking at changing it to `enforce_untyped_import_check` (/checks?) to keep default as false, but not sure if it's actually worth making the change here/if it will make more sense.

Reviewed By: ndmitchell

Differential Revision: D73868541

fbshipit-source-id: f7b91939ce85983450f25837d5ebfdf3274273ab