Factor out `find_global_references_from_definition`
Summary:
In this stack, I aim to make the bulk of global find ref computation happen in another thread, so that simple IDE services will not be blocked while an expensive find-ref is happening.

In the new world, we will split global find refs into two steps:
1. find the definition of reference: stay on the main thread
2. find references in all transitive rdeps of the definition: happen in a new thread.

This diff splits out the function for 2 to prepare for later changes.

Reviewed By: kinto0

Differential Revision: D73873698

fbshipit-source-id: 669f233c05107f72e86044198fc6fca521a25560