use contextual type hint for callable return
Summary:
This diff uses the contextual typing hint to specialize constructor calls.

This is a bit tricky to handle - the change causes a regression in tuple.rs related to whether we should use the arguments or the hint to specialize the constructed class.

In the tuple.rs test, we want the type to only be determined by the argument to `tuple()`, but in the linked github issues, we want the type to be determined by the hint.

fixes https://github.com/facebook/pyrefly/issues/140

fixes https://github.com/facebook/pyrefly/issues/96

fixes https://github.com/facebook/pyrefly/issues/92

Reviewed By: ndmitchell

Differential Revision: D73866382

fbshipit-source-id: 10c957c00a39e65241d80f7c5552146cc63dc873