Factor out some functions to prepare for running bulk of find references computation in another thread
Summary:
In this stack, I aim to make the bulk of global find ref computation happen in another thread, so that simple IDE services will not be blocked while an expensive find-ref is happening.

In the new world, we will split global find refs into two steps:
1. find the definition of reference: stay on the main thread
2. find references in all transitive rdeps of the definition: happen in a new thread.

For step 2, we need to create a new transaction from `State` separate from the transaction that's reuse for most other IDE services. We will need to populate the transaction with the latest in-memory content.

This function wraps relevant data structure in Arc and factor out some functions so that we can repopulate the transaction with latest in-memory content from another thread.

Reviewed By: kinto0

Differential Revision: D73873697

fbshipit-source-id: 20306fc03c9dd6164aaed06bb46b4d59f6f85247