Run bulk of find references computation in another thread
Summary:
With preparation from the previous diffs, this is the diff where we run bulk of find references computation in another thread, so that we can parallelize an expensive find-ref with other IDE services.

In the new world, we will split global find refs into two steps:
1. find the definition of reference: stay on the main thread
2. find references in all transitive rdeps of the definition: happen in a new thread.

For step 2, we need to create a new transaction from `State` separate from the transaction that's reuse for most other IDE services. We will need to populate the transaction with the latest in-memory content. This can be done with the functions factored out from the previous diff.

Finally, we use a thread pool of size 1 to ensure that we only have one of these expensive reads happening. It's not necessary for the architecture, but it would be good to ensure that we are not using too much memory and resources at once. (It also gives us some deterministic ordering of find-ref IDE responses, although it's also not critical since each response will have a request id anyways.)

Reviewed By: kinto0

Differential Revision: D73873696

fbshipit-source-id: 22e14634df18fe7a5d04e7bc9fad6a6480b846ff