Add implicit parent modules to site_package_path
Summary:
Otherwise we find `sys` at `/sys` as a namespace package. We want these to be findable, but less important than typeshed.

To make the IDE still work out, we need to use the configured in addition to the workspace paths. Really not sure what we should be doing here, but this is more permissive, which seems the right direction of travel.

Reviewed By: SamChou19815

Differential Revision: D73908298

fbshipit-source-id: 2ab8954d404d72a3e99729d2a1fdf3072555c13d