Make sure run state is reused between button and output pane
Summary: For the run button, we keep track of 2 isRunning state right now, one created in `MonacoEditorButton` and one created in `Sandbox`. This leads to some weird behavior where the button is processing, but the output panel already finished. This diff makes sure that we are using only one of the state for the button by passing in the existing state to `MonacoEditorButton` if it exists

Reviewed By: connernilsen

Differential Revision: D73808827

fbshipit-source-id: 0227a7592f547f1c8206b7ec2522d44c05e9e6a8