Create new constructors for workspace in lsp
Summary:
This diff changes no logic. We add new constructors for `lsp::Workspace`, and rename the old `new()` constructor to `old()`. It doesn't matter what it's named, since we're removing it in two diffs anyway.

We also add the `root` and `python_environment` fields, which will replace the `sys_info`, `search_path`, and `config_file` fields in the future. Right now, the focus is on creating/replacing the constructors, and we'll deal with removing the fields/updating logic in the future. The fields should all have consistent information (e.g. `python_environment` should match the one in `config_file`, ...), so that swapping our usage of the fields should be seamless.

Reviewed By: SamChou19815

Differential Revision: D73951016

fbshipit-source-id: 4a5b1a9b478827c701c79f519c419521d6369ada