Make the handling of attributes on type vars clearer & tested
Summary:
The `AttributeBase::Quantified` case was added to handle attributes directly
on a type var, which is important because (a) ParamSpec allows `P.args` and
`P.kwargs` and (b) invalid user code might try to access attributes of other
type vars.

I had a couple issues with this:
- The code was not documented, and the name `AttributeBase::Quantified` strongly
  suggests that it models a base which is a term with a quantified type (I found
  a TODO comment with incorrect instructions based on that faulty interpretation,
  so I don't think it's just me who was misreading this)
- There was no doc comment. It is not possible to make this code obvious, because
  it only pops up in edge cases, so it *really* needs a doc comment. I added that,
  and also put a comment in the lookup logic because someone maintaining the code
  might find that first.

There's also now a regression test of the vanilla (illegal) case where a user tries
to access a nonexistant attribute on a type var; I wasn't convinced this would
even work until I tried it.

Reviewed By: yangdanny97

Differential Revision: D73936972

fbshipit-source-id: d77a1d1cc931ada9fe0ed6a07702a94ee61f1df0