Document a bug: we cannot handle a typevar bound by a union
Summary:
Because types compose, I think it's very hard to avoid edge
cases where you have a non-union type that cannot nest a union in a
way that will break our existing layering logic in attribute access.

In this example, a type var with a union bound triggers the unknown-attr-base
failure.

This is the same architectural issue already flagged with a TODO on the
constrained type var case (which behaves the same way from an attribute
access point of view).

I suspect the long-term solution will be a refactor that removes the attempt
to break attribute access into a clearly layered api, which doesn't really work
well because of this ability to nest unions inside of other type constructs; we
really need a single, recursive layer that is able to re-enter the union logic.
It might still be possible to use two layers (separating the union case does help
a lot with readability!), but they need to be mutually recursive if so.

Reviewed By: ndmitchell

Differential Revision: D73970361

fbshipit-source-id: 88493e6f009e0db631132079cf4e8499fa07d9a1