Cancel find-ref transactions before committing
Summary:
This diff makes the first use of transaction cancellation infra so that we can cancel ongoing find-refs before committing a recheck. (If we do nothing, then a recheck commit will be blocked, because we cannot read and write to the State at the same time.)

In this diff, we will register the cancellation handle for transactions for find-ref in the State, and before committing, we will first cancel everything recorded in `cancellation_handles`.

Reviewed By: kinto0, ndmitchell

Differential Revision: D73947610

fbshipit-source-id: 7cf2e8f05e0afe56fa7fafa57c1fd110ad37382e