Support cooperative cancellation in Transaction
Summary:
Currently, we don't have cancellation support for transactions. There are two negative consequence for that:

1. A long running find-ref can block committing transaction for recheck
2. We will waste computing time when IDE decides to send $/cancel for find-ref (when another one is triggered)

This diff first adds the cooperative cancellation support in Transaction. The cooperation is inserted into the worklist: after working for each item in the list, we check whether the cancellation bool has been set to true. If so, we will cancel, and propagate the information all the way to `run`, so that externally caller can know whether a run finishes normally or is canceled in the middle.

To prevent misuse, the cancellation handle is only exposed in `CancellableTransaction`, and only the `run` method in `CancellableTransaction` will report whether the transaction finishes normally.

In the next diff, I will use the infra to implement cancellation support for find-ref.

Reviewed By: ndmitchell

Differential Revision: D73947613

fbshipit-source-id: 58d1407128d281eeeddf964aa634a9ba9889302c