Add dump-config command
Summary:
I realized while trying to debug search path issues in the run_typeshed script that it's quite difficult to figure out exactly what pyrefly is doing when it can't find an import. I suspect this will be a very common problem, since we had similar issues with pytorch and have had to make multiple search-path-related tweaks to the mypy_primer script as well.

I think it would be useful to have a command that dumps the parts of the configuration related to import resolution, both for our own sanity and to help users investigate and report issues.

Reviewed By: ndmitchell

Differential Revision: D73979622

fbshipit-source-id: 338fbbfb3cbe5686886600ff1e0a364c0a508994