Eliminate `FlowStyle::Unbound` and `FlowStyle::PossiblyUnbound`
Summary:
I noticed the existence of both `FlowStyle::Unbound` and `FlowStyle::Uninitialized`. Looking at the code, their only difference seems to be error messages. All other aspects of those two are handled rather similarly.

I don't think the difference between our `FlowStyle::Unbound` and `FlowStyle::Uninitialized` exists at runtime. The same kind of error will be raised by this snippet
```
def foo():
  if False:
    x = 0
  print(x)
```
and this snippet:
```
def foo():
  def inner():
    print(x)
  f()
  x = 0
```
and this snippet
```
def foo():
  x = 0
  del x
  print(x)
```
The name of the runtime exception for this kind of issue is called `UnboundLocalError`. But that name is slightly misleading because what the error really indicates is that the name actually exists in the function -- only that it has not been initialized with a legit value. The real difference between "unbound" and "uninitialized" IMO is that the former simply cannot find the name in scope (which corresponds to `NameError` at runtime), but we've already covered that case with the "[unknown-name]" error code.

Pyrefly currently treat the last snippet as `FlowStyle::Unbound` and the other two cases as `FlowStyle::Uninitialized`. Given that such difference does not exist at runtime, this diff proposes that we just merge the two flow styles to simplify the code.

Reviewed By: yangdanny97

Differential Revision: D74035158

fbshipit-source-id: 3793267590a2b9c71654e4617ec3385152276e6a