basic invalidation for index narrowing
Summary:
This diff will invalidate the narrowing for indexes whenever the property chain is directly assigned, just like attributes.

We add a subscript equivalent of the AssignToAttribute binding to typecheck/narrow subscripts on assignment, which replaces the old SubscriptValue binding (that only does typechecking without narrowing).

After this diff, we need to handle one more case.

When the chain contains an unknown like index `w.x[y].z`, we actually need to invalidate every index narrowing for `w.x`, since we don't know what `y` is referring to.

That will be done in another diff.

Reviewed By: stroxler

Differential Revision: D74072876

fbshipit-source-id: 295503ef15b74424d720acbe42b8294341e73965