prefix invalidation for narrowing
Summary:
If we make an assignment in a chain that contains some unknown index/slice, then we have to invalidate all the narrowing for concrete indexes of its prefix.

So if we have narrowed `x.y[0]`, `x.y[1]`, etc. Then we make an assignment to `x.y[z]` where z isn't a positive integer literal, we have to clear `x.y[0]` and `x.y[1]`. Non-index narrows under `x.y` like `x.y.z` are unaffected.

As an aside, I think the case where we have both index and attribute narrows under the same parent is going to be pretty rare. If we have some class where an index and a name refer to the same thing, then this narrowing/invalidation will be unsound. NamedTuples are probably the most common case, but they're immutable so it's OK.

This could be slightly more clever: if we're assigning to an unknown index but the type is a subtype of any existing narrows for a particular known index, then we can keep the narrowing for that index. I'm not doing it here since it doesn't sound like it would be performant.

closes https://github.com/facebook/pyrefly/issues/80

Reviewed By: stroxler

Differential Revision: D74077463

fbshipit-source-id: aba55470b8278dc35048a7b0a531c73d941dd41e