Never add the cwd to the search path if we find a pyproject.toml (#159)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/159

Previously, if we saw a pyproject.toml without a [tool.pyrefly] section, we'd first create a default ConfigFile rooted at the cwd, then add the path to the config root to the search path. But if the pyrefly section was present, we'd only add the config root. Now we add only the config root in both cases.

This also fixes a minor bug where we previously added the config root twice when the pyrefly section was present.

Reviewed By: ndmitchell

Differential Revision: D74159947

fbshipit-source-id: dcb3e55477c5d397c05645bdb974e186d074f680