add magic token to ignore part of test
Summary:
soon we will be encoding file contents into URLs (assuming someone stamps this). we want to ignore these for the tests so we don't have to update them every time they change.

this is subject to change and not done in the best way - I just want to get some quick tests working before pycon

Reviewed By: connernilsen

Differential Revision: D74085902

fbshipit-source-id: 1357e7528d7a84210be231f11b6cd34e5cd70a23