support textDocumentContentProviders in open-source extension
Summary:
VSCode supports [this](https://code.visualstudio.com/api/extension-guides/virtual-documents#textdocumentcontentprovider) DocumentContentProvider API for describing schemas to find virtual files at. Other extensions use this as a means of querying a remote cache for these (see examples in T220069278).

We want to display the builtins typeshed stubs bundled with the extension when people command+click on them: In our case, we don't have them stored anywhere except in-memory in the language server (they are compressed in the extension). the easiest way to send this content seems to be through the actual URI itself, encoded in base64. this works surprisingly well and does not require additional messages being sent back/forth with content.

This diff adds support for this provider in the external extension.

Reviewed By: SamChou19815

Differential Revision: D74086703

fbshipit-source-id: 1a1f761e1ee849f8856a92feda0b8d28ea53b79b