Mostly stop adding cwd to search path in standard_config_finder
Summary:
The guesses we add to site_package_path based on the file path are more helpful and less likely to do strange things like accidentally override parts of the stdlib.

There's still one single place where we use the default cwd search path - that's if we have a ModulePath that we can't call parent() on for some reason (config_finder.rs:70).

Reviewed By: grievejia

Differential Revision: D74192670

fbshipit-source-id: 32ca26dc51f61b2871347bc2d89f11c651d13c59